requirements.txt
Flask==3.0.3
openai==1.35.13
pandas==2.2.2
python-dotenv==1.0.1
